Some small changes
updated album's addSong and removeSong to take artist as a parameter as well then updated its respective test file. Then, I added a view class which is where the main method will be and the UI for the user to interact with. Then, I created test files for View, UserLibrary, and MusicStore, they have no tests currently as those classes are not yet finished